+++ src/v1/paths.rs
