+++ tests/integration_test.rs
@@ -1,5 +1,5 @@
-use picocadrs::assets::{Color, Model, Point3D}; // Point3D required for point macro
 use picocadrs::point;
+use picocadrs::v1::assets::{Color, Model, Point3D}; // Point3D required for point macro
 use std::ffi::OsString;
 
 #[test]
